# backend/docs/swagger.yaml
basePath: /
definitions:
...
  CreateTransactionReq:
    properties:
      amount:
        format: double
        type: number
      category_id:
        type: integer
      date:
        type: string
      description:
        type: string
      wallet_id:
        type: integer
    required:
    - amount
    - category_id
    - date
    - description
    - wallet_id
    type: object
  CreateTransactionRes:
    properties:
      category:
        $ref: '#/definitions/domain.Category'
      data:
        $ref: '#/definitions/domain.Transaction'
    required:
    - data
    type: object
...
  DeleteTransactionRes:
    properties:
      message:
        type: string
    required:
    - message
    type: object
...
  GetTransactionRes:
    properties:
      category:
        $ref: '#/definitions/domain.Category'
      data:
        $ref: '#/definitions/domain.Transaction'
    required:
    - data
    type: object
...
  UpdateTransactionReq:
    properties:
      amount:
        format: double
        type: number
      category_id:
        type: integer
      date:
        type: string
      description:
        type: string
      wallet_id:
        type: integer
    required:
    - amount
    - category_id
    - date
    - description
    - wallet_id
    type: object
  UpdateTransactionRes:
    properties:
      category:
        $ref: '#/definitions/domain.Category'
      data:
        $ref: '#/definitions/domain.Transaction'
    required:
    - data
    type: object
...
  domain.Category:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      type:
        description: income, expense
        type: string
      updated_at:
        type: string
      user_id:
        description: nil for default categories
        type: integer
    required:
    - created_at
    - id
    - name
    - type
    - updated_at
    type: object
...
  domain.Transaction:
    properties:
      amount:
        format: double
        type: number
      category_id:
        type: integer
      created_at:
        type: string
      date:
        type: string
      description:
        type: string
      id:
        type: integer
      updated_at:
        type: string
      wallet_id:
        type: integer
    required:
    - amount
    - category_id
    - created_at
    - date
    - description
    - id
    - updated_at
    - wallet_id
    type: object
...
  dto.Error:
    properties:
      code:
        type: integer
      message:
        type: string
    required:
    - code
    - message
    type: object
...
  dto.Transaction:
    properties:
      category:
        $ref: '#/definitions/domain.Category'
      data:
        $ref: '#/definitions/domain.Transaction'
    required:
    - data
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API untuk aplikasi manajemen keuangan dan pencatatan emas - ARTA
  title: ARTA API
  version: "1.0"
paths:
...
  /api/transaction:
    get:
      consumes:
      - application/json
      operationId: ListTransactions
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: wallet id
        in: query
        name: wallet_id
        required: true
        type: integer
      - description: include_category
        in: query
        name: include_category
        type: boolean
      - description: start_date
        in: query
        name: start_date
        type: string
      - description: end_date
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.Transaction'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Error'
      tags:
      - transaction
    post:
      consumes:
      - application/json
      operationId: CreateTransaction
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Unique key per submission attempt for safe retry (UUID recommended)
        in: header
        name: Idempotency-Key
        type: string
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/CreateTransactionReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/CreateTransactionRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Error'
      tags:
      - transaction
  /api/transaction/{id}:
    delete:
      consumes:
      - application/json
      operationId: DeleteTransaction
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: transaction id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/DeleteTransactionRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Error'
      tags:
      - transaction
    get:
      consumes:
      - application/json
      operationId: GetTransaction
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: transaction id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/GetTransactionRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Error'
      tags:
      - transaction
    put:
      consumes:
      - application/json
      operationId: UpdateTransaction
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: transaction id
        in: path
        name: id
        required: true
        type: integer
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/UpdateTransactionReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/UpdateTransactionRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Error'
      tags:
      - transaction
...
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: '"Type ''Bearer TOKEN''"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"