# backend/internal/feature/transaction/handler.go
package transaction

import (
	"strconv"
	"strings"
	"time"

	"github.com/artafinance/backend/internal/domain"
	"github.com/artafinance/backend/internal/dto"
	categoryfeature "github.com/artafinance/backend/internal/feature/category"
	"github.com/artafinance/backend/pkg/jwt"
	"github.com/artafinance/backend/pkg/middleware"
	"github.com/gofiber/fiber/v2"
)

// Handler exposes transaction HTTP endpoints.
type Handler struct {
	repo         *Repository
	categoryRepo *categoryfeature.Repository
	jwtMgr       *jwt.Manager
}

// NewHandler creates a new transaction handler.
func NewHandler(repo *Repository, categoryRepo *categoryfeature.Repository, jwtMgr *jwt.Manager) *Handler {
	return &Handler{repo: repo, categoryRepo: categoryRepo, jwtMgr: jwtMgr}
}

// RegisterRoutes registers transaction routes.
func (h *Handler) RegisterRoutes(router fiber.Router) {
	group := router.Group("/transaction")
	protected := group.Use(middleware.AuthMiddleware(h.jwtMgr))
	protected.Get("/", h.list)
	protected.Post("/", h.create)
	protected.Get("/:id", h.get)
	protected.Put("/:id", h.update)
	protected.Delete("/:id", h.delete)
}

// @id          ListTransactions
// @tags        transaction
// @accept      json
// @produce     json
// @param       Authorization header string true "Bearer token"
// @param       wallet_id query int true "wallet id"
// @param 			include_category query bool false "include_category"
// @param 			start_date query string false "start_date"
// @param 			end_date query string false "end_date"
// @success     200 {array} dto.Transaction
// @failure     400 {object} dto.Error
// @failure     401 {object} dto.Error
// @failure     500 {object} dto.Error
// @router      /api/transaction [get]
func (h *Handler) list(c *fiber.Ctx) error {
	walletId := c.QueryInt("wallet_id", 0)
	if walletId == 0 {
		return c.Status(fiber.StatusBadRequest).JSON(dto.Error{
			Code:    fiber.StatusBadRequest,
			Message: "wallet_id is required",
		})
	}

	includeCategory := c.QueryBool("include_category", false)
	startDateStr := c.Query("start_date", "")
	endDateStr := c.Query("end_date", "")

	var startDate, endDate time.Time
	if startDateStr != "" {
		parsedStartDate, err := time.Parse("2006-01-02", startDateStr)
		if err != nil {
			return c.Status(fiber.StatusBadRequest).JSON(dto.Error{
				Code:    fiber.StatusBadRequest,
				Message: err.Error(),
			})
		}
		startDate = parsedStartDate
	}

	if endDateStr != "" {
		parsedEndDate, err := time.Parse("2006-01-02", endDateStr)
		if err != nil {
			return c.Status(fiber.StatusBadRequest).JSON(dto.Error{
				Code:    fiber.StatusBadRequest,
				Message: err.Error(),
			})
		}
		endDate = parsedEndDate
	}

	if res, err := h.repo.GetAll(&GetAllTransactionsFilter{
		WalletId:        uint(walletId),
		IncludeCategory: includeCategory,
		StartDate:       startDate,
		EndDate:         endDate,
		OrderBy:         "date",
		OrderDirection:  "desc",
	}); err != nil {
		return c.Status(fiber.StatusInternalServerError).JSON(dto.Error{
			Code:    fiber.StatusInternalServerError,
			Message: err.Error(),
		})
	} else {
		return c.Status(fiber.StatusOK).JSON(res)
	}
}

// @id                   CreateTransaction
// @tags                 transaction
// @accept               json
// @produce              json
// @param                Authorization header string true "Bearer token"
// @param                Idempotency-Key header string false "Unique key per submission attempt for safe retry (UUID recommended)"
// @param                body body CreateTransactionReq true "body"
// @success              201 {object} CreateTransactionRes
// @failure              400 {object} dto.Error
// @failure              401 {object} dto.Error
// @failure              500 {object} dto.Error
// @router               /api/transaction [post]
func (h *Handler) create(c *fiber.Ctx) error {
	var req CreateTransactionReq
	if err := c.BodyParser(&req); err != nil {
		return c.Status(fiber.StatusBadRequest).JSON(dto.Error{Code: fiber.StatusBadRequest, Message: err.Error()})
	}

	userID := middleware.GetUserID(c)
	if userID == "" {
		return c.Status(fiber.StatusUnauthorized).JSON(dto.Error{Code: fiber.StatusUnauthorized, Message: "unauthorized"})
	}

	idempotencyKey := strings.TrimSpace(c.Get("Idempotency-Key"))
	if len(idempotencyKey) > 64 {
		return c.Status(fiber.StatusBadRequest).JSON(dto.Error{Code: fiber.StatusBadRequest, Message: "Idempotency-Key must be at most 64 characters"})
	}

	ownerID, err := h.repo.GetWalletOwnerID(req.WalletID)
	if err != nil {
		return c.Status(fiber.StatusInternalServerError).JSON(dto.Error{Code: fiber.StatusInternalServerError, Message: err.Error()})
	}
	if strconv.FormatUint(uint64(ownerID), 10) != userID {
		return c.Status(fiber.StatusUnauthorized).JSON(dto.Error{Code: fiber.StatusUnauthorized, Message: "unauthorized"})
	}

	if req.CategoryID == 0 || req.Amount <= 0 {
		return c.Status(fiber.StatusBadRequest).JSON(dto.Error{Code: fiber.StatusBadRequest, Message: "invalid category or amount"})
	}

	category, err := h.categoryRepo.GetCategoryByID(req.CategoryID)
	if err != nil {
		return c.Status(fiber.StatusBadRequest).JSON(dto.Error{Code: fiber.StatusBadRequest, Message: "invalid category"})
	}
	if category.UserID != nil && strconv.FormatUint(uint64(*category.UserID), 10) != userID {
		return c.Status(fiber.StatusUnauthorized).JSON(dto.Error{Code: fiber.StatusUnauthorized, Message: "unauthorized"})
	}

	parsedDate, err := time.Parse("2006-01-02", req.Date)
	if err != nil {
		return c.Status(fiber.StatusBadRequest).JSON(dto.Error{Code: fiber.StatusBadRequest, Message: "invalid date format"})
	}

	created, err := h.repo.CreateTransaction(&domain.Transaction{
		WalletID:    req.WalletID,
		Amount:      req.Amount,
		CategoryID:  req.CategoryID,
		Description: req.Description,
		Date:        parsedDate,
	}, idempotencyKey)
	if err != nil {
		return c.Status(fiber.StatusInternalServerError).JSON(dto.Error{Code: fiber.StatusInternalServerError, Message: err.Error()})
	}

	return c.Status(fiber.StatusCreated).JSON(CreateTransactionRes{dto.Transaction{Data: *created}})
}

// @id                   GetTransaction
// @tags                 transaction
// @accept               json
// @produce              json
// @param                Authorization header string true "Bearer token"
// @param                id path int true "transaction id"
// @success              200 {object} GetTransactionRes
// @failure              400 {object} dto.Error
// @failure              401 {object} dto.Error
// @failure              404 {object} dto.Error
// @failure              500 {object} dto.Error
// @router               /api/transaction/{id} [get]
func (h *Handler) get(c *fiber.Ctx) error {
	id, err := c.ParamsInt("id")
	if err != nil {
		return c.Status(fiber.StatusBadRequest).JSON(dto.Error{
			Code:    fiber.StatusBadRequest,
			Message: err.Error(),
		})
	}

	if transaction, err := h.repo.Get(GetFilter{
		TransactionId:   uint(id),
		IncludeCategory: true,
	}); err != nil {
		return c.Status(fiber.StatusNotFound).JSON(dto.Error{
			Code:    fiber.StatusNotFound,
			Message: err.Error(),
		})
	} else {
		return c.Status(fiber.StatusOK).JSON(GetTransactionRes{
			*transaction,
		})
	}
}

// @id                   UpdateTransaction
// @tags                 transaction
// @accept               json
// @produce              json
// @param                Authorization header string true "Bearer token"
// @param                id path int true "transaction id"
// @param                body body UpdateTransactionReq true "body"
// @success              200 {object} UpdateTransactionRes
// @failure              400 {object} dto.Error
// @failure              401 {object} dto.Error
// @failure              404 {object} dto.Error
// @failure              500 {object} dto.Error
// @router               /api/transaction/{id} [put]
func (h *Handler) update(c *fiber.Ctx) error {
	id := c.Params("id")
	var req UpdateTransactionReq
	if err := c.BodyParser(&req); err != nil {
		return c.Status(fiber.StatusBadRequest).JSON(dto.Error{Code: fiber.StatusBadRequest, Message: err.Error()})
	}

	parsedID, err := strconv.ParseUint(id, 10, 64)
	if err != nil {
		return c.Status(fiber.StatusBadRequest).JSON(dto.Error{Code: fiber.StatusBadRequest, Message: err.Error()})
	}

	tx, err := h.repo.GetTransactionByID(uint(parsedID))
	if err != nil {
		return c.Status(fiber.StatusNotFound).JSON(dto.Error{Code: fiber.StatusNotFound, Message: err.Error()})
	}

	userID := middleware.GetUserID(c)
	ownerID, err := h.repo.GetWalletOwnerID(tx.WalletID)
	if err != nil {
		return c.Status(fiber.StatusInternalServerError).JSON(dto.Error{Code: fiber.StatusInternalServerError, Message: err.Error()})
	}
	if strconv.FormatUint(uint64(ownerID), 10) != userID {
		return c.Status(fiber.StatusUnauthorized).JSON(dto.Error{Code: fiber.StatusUnauthorized, Message: "unauthorized"})
	}

	if req.WalletID != nil {
		tx.WalletID = *req.WalletID
	}
	if req.Amount != nil {
		tx.Amount = *req.Amount
	}
	if req.CategoryID != nil {
		category, err := h.categoryRepo.GetCategoryByID(*req.CategoryID)
		if err != nil {
			return c.Status(fiber.StatusBadRequest).JSON(dto.Error{Code: fiber.StatusBadRequest, Message: "invalid category"})
		}
		if category.UserID != nil && strconv.FormatUint(uint64(*category.UserID), 10) != userID {
			return c.Status(fiber.StatusUnauthorized).JSON(dto.Error{Code: fiber.StatusUnauthorized, Message: "unauthorized"})
		}
		tx.CategoryID = *req.CategoryID
	}
	if req.Description != nil {
		tx.Description = *req.Description
	}
	if req.Date != nil {
		parsedDate, err := time.Parse("2006-01-02", *req.Date)
		if err != nil {
			return c.Status(fiber.StatusBadRequest).JSON(dto.Error{Code: fiber.StatusBadRequest, Message: "invalid date format"})
		}
		tx.Date = parsedDate
	}

	updated, err := h.repo.UpdateTransaction(tx)
	if err != nil {
		return c.Status(fiber.StatusInternalServerError).JSON(dto.Error{Code: fiber.StatusInternalServerError, Message: err.Error()})
	}

	return c.Status(fiber.StatusOK).JSON(UpdateTransactionRes{dto.Transaction{Data: *updated}})
}

// @id                   DeleteTransaction
// @tags                 transaction
// @accept               json
// @produce              json
// @param                Authorization header string true "Bearer token"
// @param                id path int true "transaction id"
// @success              200 {object} DeleteTransactionRes
// @failure              400 {object} dto.Error
// @failure              401 {object} dto.Error
// @failure              404 {object} dto.Error
// @failure              500 {object} dto.Error
// @router               /api/transaction/{id} [delete]
func (h *Handler) delete(c *fiber.Ctx) error {
	id := c.Params("id")
	parsedID, err := strconv.ParseUint(id, 10, 64)
	if err != nil {
		return c.Status(fiber.StatusBadRequest).JSON(dto.Error{Code: fiber.StatusBadRequest, Message: err.Error()})
	}
	tx, err := h.repo.GetTransactionByID(uint(parsedID))
	if err != nil {
		return c.Status(fiber.StatusNotFound).JSON(dto.Error{Code: fiber.StatusNotFound, Message: err.Error()})
	}

	userID := middleware.GetUserID(c)
	ownerID, err := h.repo.GetWalletOwnerID(tx.WalletID)
	if err != nil {
		return c.Status(fiber.StatusInternalServerError).JSON(dto.Error{Code: fiber.StatusInternalServerError, Message: err.Error()})
	}
	if strconv.FormatUint(uint64(ownerID), 10) != userID {
		return c.Status(fiber.StatusUnauthorized).JSON(dto.Error{Code: fiber.StatusUnauthorized, Message: "unauthorized"})
	}

	if err := h.repo.DeleteTransaction(uint(parsedID)); err != nil {
		return c.Status(fiber.StatusInternalServerError).JSON(dto.Error{Code: fiber.StatusInternalServerError, Message: err.Error()})
	}

	return c.Status(fiber.StatusOK).JSON(DeleteTransactionRes{Message: "transaction deleted"})
}